Added detection of bundled CMake in Visual Studio
